Added a login page
By adding a login page, I was able to fix a couple errors in the
process. First, the css, img, and js folders were not allowed to serve
any files, so static file handlers were added to allow tornado to serve
them. Secondly, the bootstrap files used were not correct and replaced
with correct versions of bootstrap.