Added token revoke, xsrf forms, oauth timeout exception
When the user logs out, it will also log the user's google account out
of the app, in case they are using a public computer. XSRF protects
against forgery. OAuth timeout occurs once in a while during debugging
and shouldn't happen that much, but now there is an action that takes
place.